services/AI_analysis/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
pydantic
python-multipart

# HTTP clients
httpx
aiohttp
requests

# Document processing
PyPDF2
beautifulsoup4
lxml
pypdf
pdfminer.six
docx2txt

# Azure OpenAI and AI frameworks
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
azure-identity
openai

# Agent framework (custom - adjust if needed)
# agent-framework==1.0.0  # Uncomment if available in your registry
agent_framework

# Retry and resilience
tenacity

# Data processing
pandas

# Environment variables
python-dotenv

# Logging
structlog
